Add no_std support (#13)
* Replace imports from std:: with core::

* Move 'clap' as a dependency for the 'std' feature

Introduce a new `std` feature that can be optionally disabled for nostd
support.

The new feature is on by default, so the default behaviour of this crate
remains unchanged. Note that building with `nostd` does not yet work.

* Add a prelude to import common features from core

Some components like `TryInto`, `String, etc are not visible by default
with nostd (since they are exposed via std's prelude).

Create a crate-local prelude to expose all these.

* Use a custom error type

Drops the dependency on std::io::Error. This finally enables building
with `nostd`, which can be tested via:

    cargo build --target x86_64-unknown-uefi --no-default-features --lib

Note that trying to build the binary (e.g.: `main.rs`) with nostd is not
supported.

Co-Authored-By: Isaac Marovitz <isaacryu@icloud.com>

---------

Co-authored-by: Isaac Marovitz <isaacryu@icloud.com>